Move modules outside of packs.
