+++ app/javascript/code_snippet_events/index.js
