+++ app/javascript/concatenation/index.js
