+++ app/javascript/js_sequence_diagrams/index.js
