+++ app/javascript/jwt_generator/index.js
