+++ app/javascript/notices/index.js
