+++ app/javascript/turbolinks/index.js
