+++ app/javascript/volta_tabbed_examples/user_preference.js
